refactor(web): myapge query state 구조 분리
